DARIJA_APP/Darija_back/requirements.txt
fastapi
uvicorn
joblib
python-dotenv
numpy
boto3
google-api-python-client
pandas
pyarabic
tashaphyne
emoji
python-multipart
scikit-learn